Integrate colab stream into detect.py
