+++ utils/datasets.py
@@ -4,6 +4,8 @@ import os
 import random
 import shutil
 import time
+import base64
+import io
 from pathlib import Path
 from threading import Thread
 
@@ -13,6 +15,8 @@ import torch
 from PIL import Image, ExifTags
 from torch.utils.data import Dataset
 from tqdm import tqdm
+from IPython.display import display, Javascript
+from google.colab.output import eval_js
 
 from utils.utils import xyxy2xywh, xywh2xyxy, torch_distributed_zero_first
 
@@ -288,6 +292,169 @@ class LoadStreams:  # multiple IP or RTSP cameras
 
     def __len__(self):
         return 0  # 1E12 frames = 32 streams at 30 FPS for 30 years
+
+    
+class LoadColabStreams:
+    def __init__(self, img_size=640):
+        self.img_size = img_size
+        self.stream_data = ""
+        self.mode = ""
+        self.init_js_video()
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        js_reply = eval_js('takePhoto("{}", "{}")'.format("Capturing...", self.stream_data))
+        if not js_reply:
+            raise StopIteration
+
+        # Convert js image to array
+        jpeg_bytes = base64.b64decode(js_reply['img'].split(',')[1])
+        image_PIL = Image.open(io.BytesIO(jpeg_bytes))
+        img0 = np.array(image_PIL)
+
+        # Padded resize
+        img = letterbox(img0, new_shape=self.img_size)[0]
+
+        # Convert
+        img = img[:, :, ::-1].transpose(2, 0, 1)    # BGR to RGB, to 3x416x416
+        img = np.ascontiguousarray(img)
+
+        return "colab_stream.txt", img, img0, self
+    
+    def init_js_video(self):
+        js = Javascript('''
+                var video;
+                var div = null;
+                var stream;
+                var captureCanvas;
+                var imgElement;
+                var labelElement;
+                
+                var pendingResolve = null;
+                var shutdown = false;
+        
+                function removeDom() {
+                    stream.getVideoTracks()[0].stop();
+                    video.remove();
+                    div.remove();
+                    video = null;
+                    div = null;
+                    stream = null;
+                    imgElement = null;
+                    captureCanvas = null;
+                    labelElement = null;
+                }
+        
+                function onAnimationFrame() {
+                    if (!shutdown) {
+                        window.requestAnimationFrame(onAnimationFrame);
+                    }
+                    if (pendingResolve) {
+                        var result = "";
+                        if (!shutdown) {
+                            captureCanvas.getContext('2d').drawImage(video, 0, 0, ''' + str(self.img_size) + ''', ''' + str(self.img_size) + ''');
+                            result = captureCanvas.toDataURL('image/jpeg', 0.8)
+                        }
+                        var lp = pendingResolve;
+                        pendingResolve = null;
+                        lp(result);
+                    }
+                }
+        
+                async function createDom() {
+                    if (div !== null) {
+                        return stream;
+                    }
+
+                    div = document.createElement('div');
+                    div.style.border = '2px solid black';
+                    div.style.padding = '10px';
+                    div.style.width = '100%';
+                    div.style.maxWidth = '700px';
+                    div.style.display = 'inline-block';
+                    document.body.appendChild(div);
+                
+                    const modelOut = document.createElement('div');
+                    modelOut.innerHTML = "<span>Status:</span>";
+                    labelElement = document.createElement('span');
+                    labelElement.innerText = 'No data';
+                    labelElement.style.fontWeight = 'bold';
+                    modelOut.appendChild(labelElement);
+                    div.appendChild(modelOut);
+                    
+                    video = document.createElement('video');
+                    video.style.display = 'block';
+                    video.style.float = 'left'
+                    video.style.marginRight = '10px'
+                    video.style.marginBottom = '10px'
+                    video.width = div.clientWidth - 100;
+                    video.setAttribute('playsinline', '');
+                    video.onclick = () => { shutdown = true; };
+                    stream = await navigator.mediaDevices.getUserMedia({video: { facingMode: "environment"}});
+                    div.appendChild(video);
+
+                    imgElement = document.createElement('img');
+                    imgElement.style.position = 'absolute';
+                    imgElement.style.zIndex = 1;
+                    div.appendChild(imgElement);
+                
+                    const instruction = document.createElement('button');
+                    instruction.innerHTML = 'Stop Video'
+                    instruction.style.fontSize = 'medium'
+                    div.appendChild(instruction);
+                    instruction.onclick = () => { shutdown = true; };
+                
+                    video.srcObject = stream;
+                    await video.play();
+
+                    captureCanvas = document.createElement('canvas');
+                    captureCanvas.width = ''' + str(self.img_size) + '''; //video.videoWidth;
+                    captureCanvas.height = ''' + str(self.img_size) + '''; //video.videoHeight;
+                    window.requestAnimationFrame(onAnimationFrame);
+                
+                    return stream;
+                }
+
+                async function takePhoto(label, imgData) {
+                    if (shutdown) {
+                    removeDom();
+                    shutdown = false;
+                    return '';
+                    }
+
+                    var preCreate = Date.now();
+                    stream = await createDom();
+                    
+                    var preShow = Date.now();
+                    if (label != "") {
+                    labelElement.innerHTML = label;
+                    }
+                
+                    if (imgData != "") {
+                    var videoRect = video.getClientRects()[0];
+                    imgElement.style.top = videoRect.top + "px";
+                    imgElement.style.left = videoRect.left + "px";
+                    imgElement.style.width = videoRect.width + "px";
+                    imgElement.style.height = videoRect.height + "px";
+                    imgElement.src = imgData;
+                    }
+            
+                    var preCapture = Date.now();
+                    var result = await new Promise(function(resolve, reject) {
+                    pendingResolve = resolve;
+                    });
+                    shutdown = false;
+            
+                    return {'create': preShow - preCreate, 
+                            'show': preCapture - preShow, 
+                            'capture': Date.now() - preCapture,
+                            'img': result};
+                }
+            ''')
+
+        display(js)
 
 
 class LoadImagesAndLabels(Dataset):  # for training/testing
